+++ src/App.vue
@@ -20,9 +20,9 @@ export default {
     'footer-container': FooterContainer
   },
   mounted () { // Can't use before mount because that lifecycle isn't called if serving via static files
-    let web3 = store.get('network') ? new Web3(new Web3.providers.HttpProvider(store.get('network').url)) : new Web3(new Web3.providers.HttpProvider(this.$store.state.Networks['ETH'][0].url))
+    const notifications = store.get('notifications') !== undefined ? store.get('notifications') : {}
     const state = {
-      web3: window.web3 ? window.web3.setProvider(this.$store.state.Networks['ETH'][0].url) : web3,
+      web3: store.get('network') ? new Web3(store.get('network').url) : new Web3(this.$store.state.Networks['ETH'][0].url),
       network: store.get('network') !== undefined ? store.get('network') : this.$store.state.Networks['ETH'][0],
       wallet: null,
       account: {
@@ -37,15 +37,13 @@ export default {
           sideMenu: store.get('sideMenu') !== undefined ? store.get('sideMenu') : 'send'
         }
       },
-      notifications: store.get('notifications') !== undefined ? store.get('notifications') : {},
+      notifications: notifications,
       gasPrice: store.get('gasPrice') !== undefined ? store.get('gasPrice') : 41
     }
 
+    if (store.get('notifications') === undefined) store.set('notifications', {})
     this.$store.dispatch('setState', state)
     this.$store.dispatch('checkIfOnline')
-    if (window.web3) {
-      this.$store.dispatch('setWeb3Instance', window.web3)
-    }
   }
 }
 </script>