+++ src/helpers/index.js
@@ -2,9 +2,11 @@ import Wallet from 'ethereumjs-wallet'
 import parseTokensHex from './parseTokensHex'
 import Configs from './configs'
 import Blockies from './blockies'
+import Misc from './misc'
 export {
   Wallet,
   Configs,
   parseTokensHex,
-  Blockies
+  Blockies,
+  Misc
 }