+++ src/layouts/InterfaceLayout/containers/SendCurrencyContainer/SendCurrencyContainer.vue
@@ -101,7 +101,7 @@
       </div>
       <interface-bottom-text link="/" :linkText="$t('interface.learnMore')" :question="$t('interface.haveIssues')"></interface-bottom-text>
     </div>
-    <confirm-modal :showSuccess="showSuccessModal" :signedTx="signedTx" :fee="transactionFee" :gasPrice="$store.state.gasPrice" :from="$store.state.wallet.getAddressString()" :to="toAddress" :value="amount" :gas="gasLimit" :data="data" :nonce="nonce + 1"></confirm-modal>
+    <confirm-modal :showSuccess="showSuccessModal" :signedTx="signedTx" :fee="transactionFee" :gasPrice="$store.state.gasPrice" :from="$store.state.wallet.getAddressString()" :to="toAddress" :value="amount" :gas="gasLimit" :data="data" :nonce="nonce"></confirm-modal>
     <success-modal message="Sending Transaction" linkMessage="Close"></success-modal>
   </div>
 </template>
@@ -157,17 +157,17 @@ export default {
     showSuccessModal () {
       this.$children[6].$refs.success.show()
     },
-    createTx () {
+    async createTx () {
       const jsonInterface = [{'constant': false, 'inputs': [{'name': '_to', 'type': 'address'}, {'name': '_amount', 'type': 'uint256'}], 'name': 'transfer', 'outputs': [{'name': 'success', 'type': 'bool'}], 'payable': false, 'type': 'function'}]
       const contract = new this.$store.state.web3.eth.Contract(jsonInterface)
-      const isEth = this.selectedCurrency.name === 'Ether'
-      this.nonce = this.$store.state.web3.eth.getTransactionCount(this.$store.state.wallet.getAddressString())
+      const isEth = this.selectedCurrency.name === 'Ethereum'
+      this.nonce = await this.$store.state.web3.eth.getTransactionCount(this.$store.state.wallet.getAddressString())
       this.data = isEth ? this.data : contract.methods.transfer(this.toAddress, unit.toWei(this.amount, 'ether')).encodeABI()
 
       this.raw = {
         from: this.$store.state.wallet.getAddressString(),
         gas: this.gasLimit,
-        nonce: this.nonce + 1,
+        nonce: this.nonce,
         gasPrice: Number(unit.toWei(this.$store.state.gasPrice, 'gwei')),
         value: isEth ? this.amount === '' ? 0 : unit.toWei(this.amount, 'ether') : 0,
         to: isEth ? this.toAddress : this.selectedCurrency.addr,
@@ -198,7 +198,7 @@ export default {
       }
     },
     createDataHex () {
-      if (this.selectedCurrency.name !== 'Ether') {
+      if (this.selectedCurrency.name !== 'Ethereum') {
         const jsonInterface = [{'constant': false, 'inputs': [{'name': '_to', 'type': 'address'}, {'name': '_amount', 'type': 'uint256'}], 'name': 'transfer', 'outputs': [{'name': 'success', 'type': 'bool'}], 'payable': false, 'type': 'function'}]
         const contract = new this.$store.state.web3.eth.Contract(jsonInterface)
         this.data = contract.methods.transfer(this.toAddress, this.amount).encodeABI()