+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 ### UI/UX
 - Add Create Wallet warning page [#46](https://github.com/MyEtherWallet/MyEtherWallet/pull/46)
 
+### Feature
+- Let users deploy and/or interact with contracts [#45](https://github.com/MyEtherWallet/MyEtherWallet/pull/45)
+
+### Devop
+- Fetch tokens and contract abi on build [#41](https://github.com/MyEtherWallet/MyEtherWallet/pull/41)
+
 ### Release v0.0.2
 ### Feature
 - additional pages [#28](https://github.com/MyEtherWallet/MyEtherWallet/pull/28)